Error Cannot read property 'text' of undefined
I am using Typescript 2.1 and React combine. 
Reason: I got error when try to open index.html from exported coverage.  As I understand `structuredText[startLine]` can be null. When it is null then this code throw error `structuredText[startLine].text;.
Temp Solve : I added a control which is check `structuredText[startLine]`and if it is not null then execute inner of if blocks.